+++ src/subcommands/deploy/mod.rs
@@ -6,31 +6,18 @@ use std::path::{Path, PathBuf};
 
 use anyhow::{anyhow, Error, Result};
 use chrono::prelude::*;
-use ckb_chain_spec::consensus::TYPE_ID_CODE_HASH;
 use ckb_hash::blake2b_256;
-use ckb_index::{CellIndex, IndexDatabase, LiveCellInfo};
 use ckb_jsonrpc_types as json_types;
 use ckb_jsonrpc_types::JsonBytes;
 use ckb_sdk::{
-    constants::{ONE_CKB, SECP_SIGNATURE_SIZE},
-    Address, AddressPayload, GenesisInfo, HttpRpcClient, HumanCapacity, MultisigConfig,
-    NetworkType, SignerFn, TxHelper,
+    constants::ONE_CKB, Address, GenesisInfo, HttpRpcClient, HumanCapacity, MultisigConfig,
+    SignerFn,
 };
-use ckb_sdk_types::{
-    deployment::{
-        Cell, CellLocation, CellRecipe, DepGroup, DepGroupRecipe, Deployment, DeploymentRecipe,
-    },
-    tx_helper::{ReprMultisigConfig, ReprTxHelper},
+use ckb_sdk_types::deployment::{
+    Cell, CellLocation, CellRecipe, DepGroup, DepGroupRecipe, Deployment, DeploymentRecipe,
 };
-use ckb_types::{
-    bytes::Bytes,
-    core::{BlockView, Capacity, FeeRate, ScriptHashType, TransactionBuilder},
-    packed,
-    prelude::*,
-    H160, H256,
-};
+use ckb_types::{bytes::Bytes, core::BlockView, packed, prelude::*, H160, H256};
 use clap::{App, Arg, ArgMatches};
-use serde_derive::{Deserialize, Serialize};
 
 use super::{CliSubCommand, Output};
 use crate::plugin::PluginManager;
@@ -40,14 +27,24 @@ use crate::utils::{
         AddressParser, ArgParser, DirPathParser, FilePathParser, FixedHashParser, FromStrParser,
         PrivkeyPathParser, PrivkeyWrapper,
     },
-    index::{with_db, IndexController},
+    index::IndexController,
     other::{
-        calculate_type_id, get_keystore_signer, get_live_cell_with_cache, get_max_mature_number,
-        get_network_type, get_privkey_signer, is_mature, read_password,
+        get_keystore_signer, get_live_cell_with_cache, get_max_mature_number, get_network_type,
+        get_privkey_signer, read_password,
     },
 };
 
-const DEPLOYMENT_TOML: &str = include_str!("../deployment.toml");
+mod cell_collector;
+mod intermedium_info;
+mod state_change;
+mod tx_builder;
+
+use cell_collector::CellCollector;
+use intermedium_info::IntermediumInfo;
+use state_change::{CellChange, ChangeInfo, DepGroupChange, ReprStateChange, StateChange};
+use tx_builder::build_tx;
+
+const DEPLOYMENT_TOML: &str = include_str!("../../deployment.toml");
 const WARN_FEE_CAPACITY: u64 = ONE_CKB;
 
 // Features:
@@ -1006,284 +1003,6 @@ fn explain_txs(info: &IntermediumInfo) -> Result<()> {
     Ok(())
 }
 
-fn build_tx<T: DeployInfo>(
-    from_address: &Address,
-    collector: &mut CellCollector,
-    fee_rate: u64,
-    cell_deps: Vec<packed::CellDep>,
-    multisig_config: Option<&MultisigConfig>,
-    lock_script: &packed::Script,
-    infos: &[T],
-) -> Result<Option<packed::Transaction>> {
-    let to_capacity: u64 = infos
-        .iter()
-        .filter_map(|info| {
-            if info.has_new_output() {
-                Some(info.occupied_capacity(lock_script))
-            } else {
-                None
-            }
-        })
-        .sum();
-    if to_capacity == 0 {
-        return Ok(None);
-    }
-
-    let from_lock_hash: H256 = packed::Script::from(from_address.payload())
-        .calc_script_hash()
-        .unpack();
-    let (mut inputs, mut input_capacities): (Vec<_>, Vec<_>) =
-        infos.iter().filter_map(|info| info.build_input()).unzip();
-    if inputs.is_empty() {
-        let (more_infos, more_capacity) = collector.collect_one(from_lock_hash, true)?;
-        inputs.extend(more_infos.into_iter().map(|info| info.input()));
-        input_capacities.push(more_capacity);
-    }
-    if inputs.is_empty() {
-        return Err(anyhow!(
-            "Capacity(mature) not enough from {}, require more than {}",
-            from_address,
-            to_capacity,
-        ));
-    }
-    let first_cell_input = &inputs[0];
-    let (outputs, outputs_data): (Vec<_>, Vec<_>) = infos
-        .iter()
-        .filter_map(|info| info.build_cell_output(lock_script, first_cell_input))
-        .unzip();
-    let init_input_total_capacity: u64 = input_capacities.into_iter().sum();
-    let base_tx = TransactionBuilder::default()
-        .cell_deps(cell_deps)
-        // update witnesses for calculate transaction fee
-        .witnesses(inputs.iter().map(|_| Bytes::default().pack()))
-        .inputs(inputs)
-        .outputs(outputs)
-        .outputs_data(outputs_data.into_iter().map(|data| data.pack()))
-        .build()
-        .data();
-    let (final_tx, input_total_capacity, output_total_capacity) = tx_adjust_fee(
-        base_tx,
-        init_input_total_capacity,
-        collector,
-        fee_rate,
-        from_address.payload(),
-        multisig_config,
-    )?;
-    log::info!(
-        "transaction fee: {}",
-        HumanCapacity(input_total_capacity - output_total_capacity)
-    );
-    assert!(input_total_capacity > output_total_capacity);
-    assert!(input_total_capacity - output_total_capacity < WARN_FEE_CAPACITY);
-    Ok(Some(final_tx))
-}
-
-fn tx_fill_inputs(
-    base_tx: packed::Transaction,
-    mut input_total_capacity: u64,
-    from_lock_hash: H256,
-    collector: &mut CellCollector,
-) -> Result<(packed::Transaction, u64, u64)> {
-    let output_total_capacity: u64 = base_tx
-        .raw()
-        .outputs()
-        .into_iter()
-        .zip(base_tx.raw().outputs_data().into_iter())
-        .map(|(output, data)| {
-            output
-                .occupied_capacity(Capacity::bytes(data.len()).unwrap())
-                .unwrap()
-                .as_u64()
-        })
-        .sum();
-    let new_tx = if output_total_capacity > input_total_capacity {
-        let (more_infos, more_capacity) = collector.collect_live_cells(
-            from_lock_hash,
-            output_total_capacity - input_total_capacity,
-            true,
-        )?;
-        let more_inputs: Vec<_> = more_infos.into_iter().map(|info| info.input()).collect();
-        input_total_capacity += more_capacity;
-        base_tx
-            .as_advanced_builder()
-            // update witnesses for calculate transaction fee
-            .witnesses(more_inputs.iter().map(|_| Bytes::default().pack()))
-            .inputs(more_inputs.pack())
-            .build()
-            .data()
-    } else {
-        base_tx
-    };
-    Ok((new_tx, input_total_capacity, output_total_capacity))
-}
-
-fn tx_adjust_fee(
-    base_tx: packed::Transaction,
-    init_input_total_capacity: u64,
-    collector: &mut CellCollector,
-    fee_rate_value: u64,
-    from_address_payload: &AddressPayload,
-    multisig_config: Option<&MultisigConfig>,
-) -> Result<(packed::Transaction, u64, u64)> {
-    const MOLECULE_NUMBER_SIZE: usize = 4;
-
-    let from_lock = packed::Script::from(from_address_payload);
-    let from_lock_hash: H256 = from_lock.calc_script_hash().unpack();
-
-    let (filled_tx, input_total_capacity, output_total_capacity) = tx_fill_inputs(
-        base_tx,
-        init_input_total_capacity,
-        from_lock_hash.clone(),
-        collector,
-    )?;
-    if input_total_capacity < output_total_capacity {
-        return Err(anyhow!(
-            "Not enough capacity to build the transaction, expected more than {}, got {}",
-            HumanCapacity(output_total_capacity),
-            HumanCapacity(input_total_capacity),
-        ));
-    }
-
-    let delta_capacity = input_total_capacity - output_total_capacity;
-    let fee_rate = FeeRate::from_u64(fee_rate_value);
-    let sighash_lock_witness_size = packed::WitnessArgs::new_builder()
-        .lock(Some(Bytes::from(vec![0u8; SECP_SIGNATURE_SIZE])).pack())
-        .build()
-        .as_slice()
-        .len();
-    let multisig_lock_witness_size = multisig_config
-        .map(|config| {
-            let total_data_len = SECP_SIGNATURE_SIZE * (config.threshold() as usize)
-                + config.to_witness_data().len();
-            packed::WitnessArgs::new_builder()
-                .lock(Some(Bytes::from(vec![0u8; total_data_len])).pack())
-                .build()
-                .as_slice()
-                .len()
-        })
-        .unwrap_or_default();
-    let tx_size = filled_tx.as_reader().serialized_size_in_block()
-        + sighash_lock_witness_size
-        + multisig_lock_witness_size;
-    let min_fee = fee_rate.fee(tx_size).as_u64();
-    log::info!(
-        "input-total: {}, output-total: {}, delta-capacity: {}, tx-size: {}, fee-rate: {}, min-fee: {}",
-        HumanCapacity(input_total_capacity),
-        HumanCapacity(output_total_capacity),
-        HumanCapacity(delta_capacity),
-        tx_size,
-        fee_rate_value,
-        HumanCapacity(min_fee),
-    );
-    if min_fee == delta_capacity {
-        log::info!("transaction fee fit perfectly!");
-        return Ok((filled_tx, input_total_capacity, output_total_capacity));
-    }
-
-    let base_change_cell_output = packed::CellOutput::new_builder().lock(from_lock).build();
-    let witness_offset_length = MOLECULE_NUMBER_SIZE * 2;
-    let input_serialized_size =
-        packed::CellInput::default().as_slice().len() + witness_offset_length;
-    let change_data_offset_length = MOLECULE_NUMBER_SIZE * 2;
-    let change_output_serialized_size =
-        base_change_cell_output.as_slice().len() + MOLECULE_NUMBER_SIZE + change_data_offset_length;
-    let change_cell_occupied_capacity = base_change_cell_output
-        .occupied_capacity(Capacity::zero())
-        .unwrap()
-        .as_u64();
-
-    if collector
-        .collect_one(from_lock_hash.clone(), false)?
-        .0
-        .is_empty()
-        && min_fee > delta_capacity
-    {
-        return Err(anyhow!("No more live cells to pay transaction fee"));
-    }
-    if min_fee < delta_capacity
-        && (delta_capacity - min_fee)
-            <= fee_rate
-                .fee(input_serialized_size + change_output_serialized_size)
-                .as_u64()
-    {
-        log::info!("fee rate too high no need to adjust it by collect more inputs");
-        return Ok((filled_tx, input_total_capacity, output_total_capacity));
-    }
-
-    let mut extra_infos: Vec<LiveCellInfo> = Vec::new();
-    let mut extra_capacity: u64 = 0;
-    loop {
-        let final_tx_size =
-            tx_size + input_serialized_size * extra_infos.len() + change_output_serialized_size;
-        let final_min_fee = fee_rate.fee(final_tx_size).as_u64();
-        let final_delta_capacity = delta_capacity + extra_capacity;
-        log::info!(
-            "final-min-fee: {}, extra-capacity: {}",
-            HumanCapacity(final_min_fee),
-            HumanCapacity(extra_capacity)
-        );
-        if final_delta_capacity >= change_cell_occupied_capacity + final_min_fee {
-            let change_capacity = final_delta_capacity - final_min_fee;
-            log::info!(
-                "have enough capacity for change cell, change capacity: {}",
-                HumanCapacity(change_capacity)
-            );
-            let change_output = base_change_cell_output
-                .as_builder()
-                .capacity(Capacity::shannons(change_capacity).pack())
-                .build();
-            let final_tx = filled_tx
-                .as_advanced_builder()
-                .witnesses(extra_infos.iter().map(|_| Bytes::default().pack()))
-                .inputs(extra_infos.into_iter().map(|info| info.input()))
-                .output(change_output)
-                .output_data(Bytes::default().pack())
-                .build()
-                .data();
-            return Ok((
-                final_tx,
-                input_total_capacity + extra_capacity,
-                output_total_capacity + change_capacity,
-            ));
-        } else {
-            if extra_infos.len() >= 5 {
-                return Err(anyhow!("load >= 5 extra input cells, something is wrong!"));
-            }
-
-            log::info!("try to collect one more live cell ...");
-            let (more_infos, more_capacity) =
-                collector.collect_one(from_lock_hash.clone(), true)?;
-            if more_infos.is_empty()
-                && final_delta_capacity < change_cell_occupied_capacity + final_min_fee
-            {
-                log::info!("have no capacity for change cell");
-                if final_delta_capacity >= WARN_FEE_CAPACITY {
-                    eprintln!(
-                        "WARNING: current transaction fee = {} CKB, not enough live cell to reduce transaction fee to less than {} CKB, try to transfer some capacity to this address",
-                        HumanCapacity(final_delta_capacity),
-                        HumanCapacity(WARN_FEE_CAPACITY),
-                    );
-                }
-                // no cpacity for put the change cell
-                let final_tx = filled_tx
-                    .as_advanced_builder()
-                    .witnesses(extra_infos.iter().map(|_| Bytes::default().pack()))
-                    .inputs(extra_infos.into_iter().map(|info| info.input()))
-                    .build()
-                    .data();
-                return Ok((
-                    final_tx,
-                    input_total_capacity + extra_capacity,
-                    output_total_capacity,
-                ));
-            }
-            log::info!("collected {} more live cells", more_infos.len());
-            extra_infos.extend(more_infos);
-            extra_capacity += more_capacity;
-        }
-    }
-}
-
 fn load_input_txs(
     input_txs: &mut HashMap<H256, json_types::Transaction>,
     rpc_client: &mut HttpRpcClient,
@@ -1310,611 +1029,6 @@ fn load_input_txs(
         input_txs.insert(tx_hash, input_tx);
     }
     Ok(())
-}
-
-#[derive(Clone, Default, Debug, Serialize, Deserialize)]
-struct IntermediumInfo {
-    deployment: Deployment,
-    last_recipe: Option<DeploymentRecipe>,
-    new_recipe: DeploymentRecipe,
-    // For offline sign (should verify the tx hash)
-    used_input_txs: HashMap<H256, json_types::Transaction>,
-    cell_tx: Option<json_types::Transaction>,
-    cell_tx_signatures: HashMap<JsonBytes, Vec<JsonBytes>>,
-    cell_changes: Vec<ReprStateChange>,
-    dep_group_tx: Option<json_types::Transaction>,
-    dep_group_tx_signatures: HashMap<JsonBytes, Vec<JsonBytes>>,
-    dep_group_changes: Vec<ReprStateChange>,
-}
-
-impl IntermediumInfo {
-    fn multisig_configs(&self) -> Result<HashMap<H160, ReprMultisigConfig>> {
-        // NOTE: we don't care the NetworkType here.
-        let network = NetworkType::Testnet;
-        let mut multisig_configs = HashMap::default();
-        if !self.deployment.multisig_config.sighash_addresses.is_empty() {
-            let config = MultisigConfig::try_from(self.deployment.multisig_config.clone())
-                .map_err(Error::msg)?;
-            multisig_configs.insert(config.hash160(), config.into_repr(network));
-        }
-        Ok(multisig_configs)
-    }
-
-    fn cell_tx_helper(&self) -> Result<Option<TxHelper>> {
-        if let Some(cell_tx) = self.cell_tx.as_ref() {
-            let repr = ReprTxHelper {
-                transaction: cell_tx.clone(),
-                multisig_configs: self.multisig_configs()?,
-                signatures: self.cell_tx_signatures.clone(),
-            };
-            let helper = TxHelper::try_from(repr).map_err(Error::msg)?;
-            Ok(Some(helper))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn dep_group_tx_helper(&self) -> Result<Option<TxHelper>> {
-        if let Some(dep_group_tx) = self.dep_group_tx.as_ref() {
-            let repr = ReprTxHelper {
-                transaction: dep_group_tx.clone(),
-                multisig_configs: self.multisig_configs()?,
-                signatures: self.dep_group_tx_signatures.clone(),
-            };
-            let helper = TxHelper::try_from(repr).map_err(Error::msg)?;
-            Ok(Some(helper))
-        } else {
-            Ok(None)
-        }
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-enum StateChange<C, R> {
-    Changed {
-        // New data
-        data: Bytes,
-        data_hash: H256,
-        config: C,
-        old_recipe: R,
-        output_index: u64,
-    },
-    NewAdded {
-        data: Bytes,
-        data_hash: H256,
-        config: C,
-        output_index: u64,
-    },
-    Unchanged {
-        data: Bytes,
-        data_hash: H256,
-        config: C,
-        old_recipe: R,
-    },
-    Removed {
-        old_recipe: R,
-    },
-}
-
-impl<C, R> StateChange<C, R> {
-    fn has_new_output(&self) -> bool {
-        match self {
-            StateChange::Changed { .. } => true,
-            StateChange::NewAdded { .. } => true,
-            StateChange::Removed { .. } => false,
-            StateChange::Unchanged { .. } => false,
-        }
-    }
-}
-
-#[derive(Clone, Default, Debug, Serialize, Deserialize)]
-struct ReprStateChange {
-    name: String,
-    kind: String,
-    old_capacity: u64,
-    new_capacity: u64,
-}
-
-type CellChange = StateChange<Cell, CellRecipe>;
-type DepGroupChange = StateChange<DepGroup, DepGroupRecipe>;
-
-trait DeployInfo {
-    fn name(&self) -> &String;
-    fn to_repr(&self, lock_script: &packed::Script) -> ReprStateChange;
-    fn has_new_output(&self) -> bool;
-    fn has_new_recipe(&self) -> bool;
-    fn occupied_capacity(&self, lock_script: &packed::Script) -> u64;
-    fn build_input(&self) -> Option<(packed::CellInput, u64)>;
-    fn build_cell_output(
-        &self,
-        lock_script: &packed::Script,
-        first_cell_input: &packed::CellInput,
-    ) -> Option<(packed::CellOutput, Bytes)>;
-}
-
-impl DeployInfo for CellChange {
-    fn name(&self) -> &String {
-        match self {
-            StateChange::Changed { config, .. } => &config.name,
-            StateChange::NewAdded { config, .. } => &config.name,
-            StateChange::Unchanged { config, .. } => &config.name,
-            StateChange::Removed { old_recipe } => &old_recipe.name,
-        }
-    }
-
-    fn to_repr(&self, lock_script: &packed::Script) -> ReprStateChange {
-        let new_capacity = self.occupied_capacity(lock_script);
-        let (kind, old_capacity) = match self {
-            StateChange::Changed { old_recipe, .. } => ("Changed", old_recipe.occupied_capacity),
-            StateChange::NewAdded { .. } => ("NewAdded", 0),
-            StateChange::Unchanged { .. } => ("Unchanged", new_capacity),
-            StateChange::Removed { old_recipe } => ("Removed", old_recipe.occupied_capacity),
-        };
-        ReprStateChange {
-            name: self.name().clone(),
-            kind: kind.to_string(),
-            old_capacity,
-            new_capacity,
-        }
-    }
-
-    fn has_new_output(&self) -> bool {
-        StateChange::has_new_output(self)
-    }
-
-    fn has_new_recipe(&self) -> bool {
-        match self {
-            StateChange::Removed { .. } => false,
-            _ => true,
-        }
-    }
-
-    fn occupied_capacity(&self, lock_script: &packed::Script) -> u64 {
-        let (data, config) = match self {
-            StateChange::Removed { .. } => return 0,
-            StateChange::Changed { data, config, .. } => (data, config),
-            StateChange::Unchanged { data, config, .. } => (data, config),
-            StateChange::NewAdded { data, config, .. } => (data, config),
-        };
-        let data_size = data.len() as u64;
-        let type_script_size: u64 = if config.enable_type_id {
-            32 + 1 + 32
-        } else {
-            0
-        };
-        lock_script.occupied_capacity().expect("capacity").as_u64()
-            + (type_script_size + data_size + 8) * ONE_CKB
-    }
-
-    fn build_input(&self) -> Option<(packed::CellInput, u64)> {
-        match self {
-            StateChange::Changed { old_recipe, .. } => {
-                let out_point = packed::OutPoint::new(old_recipe.tx_hash.pack(), old_recipe.index);
-                let input = packed::CellInput::new(out_point, 0);
-                Some((input, old_recipe.occupied_capacity))
-            }
-            _ => None,
-        }
-    }
-
-    fn build_cell_output(
-        &self,
-        lock_script: &packed::Script,
-        first_cell_input: &packed::CellInput,
-    ) -> Option<(packed::CellOutput, Bytes)> {
-        let (data, config, output_index, old_type_id) = match self {
-            StateChange::Removed { .. } => {
-                return None;
-            }
-            StateChange::Unchanged { .. } => {
-                return None;
-            }
-            StateChange::Changed {
-                data,
-                config,
-                old_recipe,
-                output_index,
-                ..
-            } => (data, config, *output_index, old_recipe.type_id.clone()),
-            StateChange::NewAdded {
-                data,
-                config,
-                output_index,
-                ..
-            } => (data, config, *output_index, None),
-        };
-        let type_id = if config.enable_type_id {
-            old_type_id.or_else(|| {
-                Some(H256::from(calculate_type_id(
-                    first_cell_input,
-                    output_index,
-                )))
-            })
-        } else {
-            None
-        };
-        let occupied_capacity = self.occupied_capacity(lock_script);
-        let type_script_opt = type_id.map(|type_id_args| {
-            packed::Script::new_builder()
-                .code_hash(TYPE_ID_CODE_HASH.pack())
-                .hash_type(ScriptHashType::Type.into())
-                .args(Bytes::from(type_id_args.as_bytes().to_vec()).pack())
-                .build()
-        });
-        let output = packed::CellOutput::new_builder()
-            .capacity(Capacity::shannons(occupied_capacity).pack())
-            .lock(lock_script.clone())
-            .type_(
-                packed::ScriptOpt::new_builder()
-                    .set(type_script_opt)
-                    .build(),
-            )
-            .build();
-        Some((output, data.clone()))
-    }
-}
-
-impl CellChange {
-    fn build_new_recipe(
-        &self,
-        lock_script: &packed::Script,
-        first_cell_input: &packed::CellInput,
-        new_tx_hash: &H256,
-    ) -> Option<CellRecipe> {
-        let (tx_hash, index, data_hash, config, old_type_id) = match self {
-            StateChange::Removed { .. } => {
-                return None;
-            }
-            StateChange::Changed {
-                data_hash,
-                config,
-                old_recipe,
-                output_index,
-                ..
-            } => (
-                new_tx_hash.clone(),
-                *output_index as u32,
-                data_hash.clone(),
-                config,
-                old_recipe.type_id.clone(),
-            ),
-            StateChange::Unchanged {
-                data_hash,
-                config,
-                old_recipe,
-                ..
-            } => (
-                old_recipe.tx_hash.clone(),
-                old_recipe.index,
-                data_hash.clone(),
-                config,
-                old_recipe.type_id.clone(),
-            ),
-            StateChange::NewAdded {
-                data_hash,
-                config,
-                output_index,
-                ..
-            } => (
-                new_tx_hash.clone(),
-                *output_index as u32,
-                data_hash.clone(),
-                config,
-                None,
-            ),
-        };
-        let type_id = if config.enable_type_id {
-            old_type_id.or_else(|| {
-                Some(H256::from(calculate_type_id(
-                    first_cell_input,
-                    index as u64,
-                )))
-            })
-        } else {
-            None
-        };
-        Some(CellRecipe {
-            name: self.name().clone(),
-            // To be replaced with final transaction hash
-            tx_hash,
-            index,
-            occupied_capacity: self.occupied_capacity(lock_script),
-            data_hash,
-            type_id,
-        })
-    }
-}
-
-impl DeployInfo for DepGroupChange {
-    fn name(&self) -> &String {
-        match self {
-            StateChange::Changed { config, .. } => &config.name,
-            StateChange::NewAdded { config, .. } => &config.name,
-            StateChange::Unchanged { config, .. } => &config.name,
-            StateChange::Removed { old_recipe } => &old_recipe.name,
-        }
-    }
-
-    fn to_repr(&self, lock_script: &packed::Script) -> ReprStateChange {
-        let new_capacity = self.occupied_capacity(lock_script);
-        let (kind, old_capacity) = match self {
-            StateChange::Changed { old_recipe, .. } => ("Changed", old_recipe.occupied_capacity),
-            StateChange::NewAdded { .. } => ("NewAdded", 0),
-            StateChange::Unchanged { .. } => ("Unchanged", new_capacity),
-            StateChange::Removed { old_recipe } => ("Removed", old_recipe.occupied_capacity),
-        };
-        ReprStateChange {
-            name: self.name().clone(),
-            kind: kind.to_string(),
-            old_capacity,
-            new_capacity,
-        }
-    }
-
-    fn has_new_output(&self) -> bool {
-        StateChange::has_new_output(self)
-    }
-
-    fn has_new_recipe(&self) -> bool {
-        match self {
-            StateChange::Removed { .. } => false,
-            _ => true,
-        }
-    }
-
-    fn occupied_capacity(&self, lock_script: &packed::Script) -> u64 {
-        let data = match self {
-            StateChange::Removed { .. } => return 0,
-            StateChange::Changed { data, .. } => data,
-            StateChange::Unchanged { data, .. } => data,
-            StateChange::NewAdded { data, .. } => data,
-        };
-        let data_size = data.len() as u64;
-        lock_script.occupied_capacity().expect("capacity").as_u64() + (data_size + 8) * ONE_CKB
-    }
-
-    fn build_input(&self) -> Option<(packed::CellInput, u64)> {
-        match self {
-            StateChange::Changed { old_recipe, .. } => {
-                let out_point = packed::OutPoint::new(old_recipe.tx_hash.pack(), old_recipe.index);
-                let input = packed::CellInput::new(out_point, 0);
-                Some((input, old_recipe.occupied_capacity))
-            }
-            _ => None,
-        }
-    }
-
-    fn build_cell_output(
-        &self,
-        lock_script: &packed::Script,
-        _first_cell_input: &packed::CellInput,
-    ) -> Option<(packed::CellOutput, Bytes)> {
-        let data = match self {
-            StateChange::Removed { .. } => {
-                return None;
-            }
-            StateChange::Unchanged { .. } => {
-                return None;
-            }
-            StateChange::Changed { data, .. } => data,
-            StateChange::NewAdded { data, .. } => data,
-        };
-        let occupied_capacity = self.occupied_capacity(lock_script);
-        let output = packed::CellOutput::new_builder()
-            .capacity(Capacity::shannons(occupied_capacity).pack())
-            .lock(lock_script.clone())
-            .build();
-        Some((output, data.clone()))
-    }
-}
-
-impl DepGroupChange {
-    fn build_new_recipe(
-        &self,
-        lock_script: &packed::Script,
-        new_tx_hash: H256,
-    ) -> Option<DepGroupRecipe> {
-        let (tx_hash, index, data_hash) = match self {
-            StateChange::Removed { .. } => {
-                return None;
-            }
-            StateChange::Changed {
-                data_hash,
-                output_index,
-                ..
-            } => (new_tx_hash, *output_index as u32, data_hash.clone()),
-            StateChange::Unchanged {
-                data_hash,
-                old_recipe,
-                ..
-            } => (
-                old_recipe.tx_hash.clone(),
-                old_recipe.index,
-                data_hash.clone(),
-            ),
-            StateChange::NewAdded {
-                data_hash,
-                output_index,
-                ..
-            } => (new_tx_hash, *output_index as u32, data_hash.clone()),
-        };
-        Some(DepGroupRecipe {
-            name: self.name().clone(),
-            // To be replaced with final transaction hash
-            tx_hash,
-            index,
-            data_hash,
-            occupied_capacity: self.occupied_capacity(lock_script),
-        })
-    }
-}
-
-struct CellCollector<'a> {
-    rpc_client: &'a mut HttpRpcClient,
-    genesis_info: &'a GenesisInfo,
-    index_dir: &'a PathBuf,
-    index_controller: IndexController,
-    wait_for_sync: bool,
-    max_mature_number: u64,
-    locked_cells: HashSet<(H256, u32)>,
-    // lock_hash => live cell
-    offchain_live_cells: Vec<LiveCellInfo>,
-}
-
-impl<'a> CellCollector<'a> {
-    pub fn new(
-        rpc_client: &'a mut HttpRpcClient,
-        genesis_info: &'a GenesisInfo,
-        index_dir: &'a PathBuf,
-        index_controller: IndexController,
-        wait_for_sync: bool,
-        max_mature_number: u64,
-    ) -> CellCollector<'a> {
-        CellCollector {
-            rpc_client,
-            genesis_info,
-            index_dir,
-            index_controller,
-            wait_for_sync,
-            max_mature_number,
-            locked_cells: Default::default(),
-            offchain_live_cells: Default::default(),
-        }
-    }
-
-    pub fn lock_cell(&mut self, tx_hash: H256, index: u32) {
-        log::debug!("lock cell, tx_hash: {:#x}, index: {}", tx_hash, index);
-        self.locked_cells.insert((tx_hash, index));
-    }
-    pub fn apply_tx(&mut self, tx: packed::Transaction) {
-        let tx_view = tx.into_view();
-        let tx_hash: H256 = tx_view.hash().unpack();
-        log::debug!("apply transaction to cell collector: {:#x}", tx_hash);
-        for out_point in tx_view.input_pts_iter() {
-            self.lock_cell(out_point.tx_hash().unpack(), out_point.index().unpack());
-        }
-        for (output_index, (output, data)) in tx_view.outputs_with_data_iter().enumerate() {
-            let type_hashes = output.type_().to_opt().map(|_| Default::default());
-            let capacity: u64 = output.capacity().unpack();
-            log::debug!(
-                "manual add live cell, tx-hash: {:#x}, index: {}, capacity: {}",
-                tx_hash,
-                output_index,
-                HumanCapacity(capacity)
-            );
-            let info = LiveCellInfo {
-                tx_hash: tx_hash.clone(),
-                output_index: output_index as u32,
-                data_bytes: data.len() as u64,
-                type_hashes,
-                lock_hash: output.lock().calc_script_hash().unpack(),
-                capacity,
-                number: Default::default(),
-                index: CellIndex {
-                    tx_index: Default::default(),
-                    output_index: output_index as u32,
-                },
-            };
-            self.offchain_live_cells.push(info);
-        }
-    }
-
-    pub fn collect_one(
-        &mut self,
-        lock_hash: H256,
-        apply_changes: bool,
-    ) -> Result<(Vec<LiveCellInfo>, u64)> {
-        self.collect_live_cells(lock_hash, 1, apply_changes)
-    }
-
-    pub fn collect_live_cells(
-        &mut self,
-        lock_hash: H256,
-        capacity: u64,
-        apply_changes: bool,
-    ) -> Result<(Vec<LiveCellInfo>, u64)> {
-        fn enough_capacity(from_capacity: u64, to_capacity: u64) -> bool {
-            from_capacity >= to_capacity
-        }
-
-        let mut collected_capacity = 0;
-        let (mut infos, rest_infos): (Vec<_>, Vec<_>) = self
-            .offchain_live_cells
-            .clone()
-            .into_iter()
-            .partition(|info| {
-                if enough_capacity(collected_capacity, capacity) {
-                    false
-                } else if info.lock_hash == lock_hash
-                    && info.type_hashes.is_none()
-                    && info.data_bytes == 0
-                {
-                    log::debug!(
-                        "got offchain live cell tx-hash: {:#x}, index: {}",
-                        info.tx_hash,
-                        info.output_index
-                    );
-                    collected_capacity += info.capacity;
-                    true
-                } else {
-                    log::debug!(
-                        "skip offchain live cell tx-hash: {:#x}, index: {}",
-                        info.tx_hash,
-                        info.output_index
-                    );
-                    false
-                }
-            });
-        if apply_changes {
-            self.offchain_live_cells = rest_infos;
-        }
-        if enough_capacity(collected_capacity, capacity) {
-            return Ok((infos, collected_capacity));
-        }
-
-        let max_mature_number: u64 = self.max_mature_number;
-        let locked_cells = self.locked_cells.clone();
-        let mut terminator = |_, info: &LiveCellInfo| {
-            if locked_cells.contains(&(info.tx_hash.clone(), info.output_index)) {
-                log::debug!(
-                    "skip locked live cell tx-hash: {:#x}, index: {}",
-                    info.tx_hash,
-                    info.output_index
-                );
-            }
-            if enough_capacity(collected_capacity, capacity) {
-                (true, false)
-            } else if info.type_hashes.is_none()
-                && info.data_bytes == 0
-                && is_mature(info, max_mature_number)
-                && !locked_cells.contains(&(info.tx_hash.clone(), info.output_index))
-            {
-                collected_capacity += info.capacity;
-                (enough_capacity(collected_capacity, capacity), true)
-            } else {
-                (false, false)
-            }
-        };
-
-        let func =
-            |db: IndexDatabase| db.get_live_cells_by_lock(lock_hash.pack(), None, &mut terminator);
-        let more_infos = with_db(
-            func,
-            self.rpc_client,
-            self.genesis_info.clone(),
-            self.index_dir,
-            self.index_controller.clone(),
-            self.wait_for_sync,
-        )
-        .map_err(Error::msg)?;
-        infos.extend(more_infos);
-        if apply_changes {
-            for info in &infos {
-                self.lock_cell(info.tx_hash.clone(), info.output_index);
-            }
-        }
-        Ok((infos, collected_capacity))
-    }
 }
 
 fn modify_info_file<T, F: FnOnce(&mut IntermediumInfo) -> Result<T>>(
